{{ cookiecutter.folder_name }}/requirements.txt
click
pytest
coverage
awscli
flake8
numpy
pandas
jupyter
pyyaml
coloredlogs
versioneer
matplotlib
